Implement mostRecentVersionOrder
