+++ apps/spruce/src/gql/queries/waterfall.graphql
@@ -31,6 +31,7 @@ query Waterfall($options: WaterfallOptions!) {
     pagination {
       hasNextPage
       hasPrevPage
+      mostRecentVersionOrder
       nextPageOrder
       prevPageOrder
     }
