+++ apps/spruce/src/pages/waterfall/testData.ts
@@ -231,3 +231,141 @@ export const getTaskStatsMock = (
     },
   },
 });
+
+export const versions: Version[] = [
+  {
+    id: "a",
+    author: "sophie.stadler",
+    activated: false,
+    createTime: new Date("2024-09-20T14:56:08Z"),
+    errors: [],
+    message: "bar",
+    requester: "gitter_request",
+    revision: "a",
+    order: 5,
+    waterfallBuilds: null,
+  },
+  {
+    id: "b",
+    author: "sophie.stadler",
+    activated: true,
+    createTime: new Date("2024-09-19T14:56:08Z"),
+    errors: [],
+    message: "foo",
+    requester: "gitter_request",
+    revision: "b",
+    order: 4,
+    waterfallBuilds: [
+      {
+        id: "ii",
+        buildVariant: "1",
+        displayName: "BV 1",
+        tasks: [
+          {
+            displayName: "Task 20",
+            displayStatusCache: "started",
+            execution: 0,
+            id: "task_20",
+            status: "started",
+          },
+          {
+            displayName: "Task 15",
+            displayStatusCache: "started",
+            execution: 0,
+            id: "task_15",
+            status: "started",
+          },
+        ],
+      },
+      {
+        id: "ii2",
+        buildVariant: "2",
+        displayName: "BV 2",
+        tasks: [
+          {
+            displayName: "Task 100",
+            displayStatusCache: "started",
+            execution: 0,
+            id: "task_100",
+            status: "started",
+          },
+        ],
+      },
+    ],
+  },
+  {
+    id: "c",
+    author: "sophie.stadler",
+    activated: true,
+    createTime: new Date("2024-09-19T14:56:08Z"),
+    errors: [],
+    message: "foo",
+    requester: "gitter_request",
+    revision: "c",
+    order: 3,
+    waterfallBuilds: [
+      {
+        id: "iii",
+        buildVariant: "3",
+        displayName: "BV 3",
+        tasks: [
+          {
+            displayName: "Task 1",
+            displayStatusCache: "",
+            execution: 0,
+            id: "task_1",
+            status: "success",
+          },
+          {
+            displayName: "Task 2",
+            displayStatusCache: "task-timed-out",
+            execution: 0,
+            id: "task_2",
+            status: "failed",
+          },
+        ],
+      },
+    ],
+  },
+  {
+    id: "d",
+    author: "sophie.stadler",
+    activated: false,
+    createTime: new Date("2024-09-19T14:56:08Z"),
+    errors: [],
+    message: "foo",
+    requester: "gitter_request",
+    revision: "d",
+    order: 2,
+  },
+  {
+    id: "e",
+    author: "sophie.stadler",
+    activated: false,
+    createTime: new Date("2024-09-19T14:56:08Z"),
+    errors: [],
+    message: "foo",
+    requester: "gitter_request",
+    revision: "e",
+    order: 1,
+  },
+  {
+    id: "f",
+    author: "sophie.stadler",
+    activated: true,
+    createTime: new Date("2024-09-19T14:56:08Z"),
+    errors: [],
+    message: "foo",
+    requester: "gitter_request",
+    revision: "f",
+    order: 0,
+    waterfallBuilds: [
+      {
+        id: "i",
+        buildVariant: "1",
+        displayName: "BV 1",
+        tasks: [],
+      },
+    ],
+  },
+];
